backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
